refactor: Rename `GlobalRightClickHandler` to `GameServerOverviewPageRightClickHandler`
Updated the component name, file paths, and imports to align with usage context and improve clarity.
