+++ src/components/display/GameServer/GameServerHouse/GameServerHouse.tsx
@@ -1,6 +1,5 @@
 import RightClickMenu from "@components/display/configurations/RightClickMenu/RightClickMenu.tsx";
-import { DeleteGameServerAlertDialog } from "@components/display/GameServerConfiguration/DeleteGameServerAlertDialog/DeleteGameServerAlertDialog.tsx";
-import GameSign from "@components/display/GameServerConfiguration/GameSign/GameSign.tsx";
+import { DeleteGameServerAlertDialog } from "@components/display/GameServer/DeleteGameServerAlertDialog/DeleteGameServerAlertDialog.tsx";
 import Link from "@components/ui/Link.tsx";
 import type { CSSProperties } from "react";
 import { useState } from "react";
@@ -10,6 +9,7 @@ import type { GameServerConfigurationEntity } from "@/api/generated/model";
 import serverHouseImage from "@/assets/ai-generated/main-page/house.png";
 import useDataInteractions from "@/hooks/useDataInteractions/useDataInteractions.tsx";
 import { cn } from "@/lib/utils.ts";
+import GameSign from "../GameSign/GameSign";
 
 const GameServerHouse = (props: {
   gameServer: GameServerConfigurationEntity;
