+++ src/i18n/de-DE/translation.ts
@@ -21,7 +21,7 @@ const translation: i18nLanguage = {
     title: "Möchten Sie {{serverName}} wirklich löschen?",
     description:
       "Diese Aktion kann nicht rückgängig gemacht werden. Dadurch wird Ihre Gameserver-Konfiguration dauerhaft gelöscht und die zugehörigen Daten von dem Server entfernt.",
-    inputLabel: "Geben Sie den Namen des Gameservers ein, um sicherzugehen: ({{serverName}})",
+    inputLabel: "Geben Sie den Namen des Gameservers ein, um sicherzugehen:",
     cancel: "Abbrechen",
     confirm: "Löschen",
   },
